requirements.txt
fastapi
openai
uvicorn
pydantic
python-dotenv
uuid
google-api-python-client
google-generativeai
google-genai
python-multipart
anthropic
Groq